another try

+++ KonaAnalyzer/Models/DayChange.cs
@@ -22,7 +22,7 @@ namespace KonaAnalyzer.Models
     {
         public DateTime Date { get; set; }
 
-        public int Change { get; set; }
+        public double Change { get; set; }
 
         //[PrimaryKey,AutoIncrement]
         //public int Id { get; set; }
